refactor: 분석 리포트 페이지 리팩토링
